requirements.txt
streamlit
pandas
google-generativeai
openpyxl
pdfplumber
pypdf